Require 'construi' in spec_helper
This gives more accurate, project wide, code coverage info.

Refactor Construi::Runner to it's own file
