+++ spec/spec_helper.rb
@@ -14,6 +14,8 @@ SimpleCov.start do
   add_filter '/spec'
 end
 
+require 'construi'
+
 RSpec.configure do |config|
 
   # rspec-expectations config goes here. You can use an alternate
